Split `Header.tsx` into multiple files
